cambio de asignacion a los campos de los numeros ya que el tamanio era demasiado chico para el dni y el numero de celular

+++ database/migrations/2023_03_01_000100_create_users_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->boolean('active')->default(true);
             $table->string('name', 100);
             $table->string('lastname', 100);
-            $table->unsignedInteger('dni')->unique();
-            $table->unsignedInteger('phone')->nullable(false);
+            $table->unsignedBigInteger('dni')->unique();
+            $table->unsignedBigInteger('phone')->nullable(false);
             $table->timestamp('phone_verified_at')->nullable();
             $table->string('password', 50);
             $table->rememberToken();
